Run golangci-lint on save only
Since golangci-lint checks original files, we can only use this checker if the
buffer is actually saved and is equal to the disk file. I don't think any reason
exists to run golangci-lint on every change, because it'll show us outdated
info, because not saved buffer differs from its on disk file.